+++ dotnet-roslyn/analyzers/Audacia.CodeAnalysis.Analyzers/src/Audacia.CodeAnalysis.Analyzers/Common/DiagnosticId.cs
@@ -27,5 +27,6 @@
         public const string TestNamesMustBeConsistent = "ACL1022";
         public const string LogMessagesNamedPropertiesPascalCase = "ACL1023";
         public const string LogMessagesNoDuplicateParameters = "ACL1024";
+        public const string LogMessagesNoExceptionsAsTemplateParameters = "ACL1025";
     }
 }